Add config to be notified on incompatible versions
Now that the updater by default only notifies on compatible versions, give
the user the option to be notified on incompatible versions (e.g. newer
versions for a newer version of Minecraft) as well. This way the user
knows when it can be safe to update to a recent version of Minecraft which
will work with this mod.